pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Alexey Kim <kmlx@yandex.ru>"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.dependencies]
python = "^3.8"
flake8 = "^3.9.2"
PyYAML = "^5.4.1"


[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"
pytest = "^6.2.3"
pytest-cov = "^2.11.1"
argparse = "^1.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
